Restructure for object construction
Kept within an isolated folder to be easier to find for individuals new to the project.
