+++ src/main/java/component_construction/ParameterError.java
@@ -1,4 +1,4 @@
-package builders;
+package component_construction;
 
 public class ParameterError extends RuntimeException {
     public ParameterError(String message) {
