+++ src/main/java/component_construction/factories/code_generation/CodeGeneratorFactory.java
@@ -1,4 +1,4 @@
-package builders;
+package component_construction.factories.code_generation;
 
 import code_generation.CodeGenerator;
 import grammar_objects.RuleConvertor;
