+++ src/main/java/component_construction/factories/lexical_analysis/LexicalAnalyserFactory.java
@@ -1,4 +1,4 @@
-package builders;
+package component_construction.factories.lexical_analysis;
 
 import lexical_analysis.*;
