+++ src/main/java/component_construction/factories/syntax_analysis/CLR1ParserFactory.java
@@ -1,6 +1,5 @@
-package builders.concrete_factories;
+package component_construction.factories.syntax_analysis;
 
-import builders.SyntaxAnalyserFactory;
 import grammar_objects.GrammarParts;
 import syntax_analysis.*;
