+++ src/main/java/component_construction/factories/syntax_analysis/SyntaxAnalyserFactory.java
@@ -1,4 +1,4 @@
-package builders;
+package component_construction.factories.syntax_analysis;
 
 import grammar_objects.*;
 import syntax_analysis.SyntaxAnalyser;
