+++ src/test/java/component_construction/builders/CompilerBuilderTests.java
@@ -1,9 +1,12 @@
-package builders;
+package component_construction.builders;
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
-import builders.concrete_factories.*;
+import component_construction.ParameterError;
+import component_construction.factories.code_generation.BasicCodeGenFactory;
+import component_construction.factories.lexical_analysis.GeneralLexicalAnalyserFactory;
+import component_construction.factories.syntax_analysis.CLR1ParserFactory;
 import grammars.basic_identifier.BasicIdentifierGrammar;
 import grammars.basic_identifier.convertors.XToYToXSemantic;
 import lexical_analysis.DynamicTokenRegex;
